+++ NEWS.md
@@ -1,3 +1,8 @@
+## dbpedia v0.1.1.9008
+
+* Error avoided when `get_dbpedia_uris()` is restricted to pre-annotated
+named entities and the document does not contain any (addresses issue #23)
+
 ## dbpedia v0.1.1.9006
 
 * Error avoided when `get_dbpedia_uris()` does not detect any URI.